refactor(config_flow): split monolithic 967-line module into a per-transport package (#354)
`config_flow.py` grew to ~1000 lines handling four transports (cloud OAuth,
cloud user-account, Modbus-only, retired cloud+modbus) plus reauth,
reconfigure, options, zeroconf discovery, and the manual-code path. Editing
any one transport risked touching another. Split into a `config_flow/`
package so each transport lives in its own module and the mental map is
smaller per file:

```
config_flow/
├── __init__.py         SungrowConfigFlow shell + async_step_user + cloud_credentials
├── _base.py            _SungrowFlowBase (state, async_remove, shared library imports)
├── _helpers.py         _normalize_redirect_uri, _parse_winet_properties,
│                       _parse_extra_measure_points, CALLBACK_WAIT_TIMEOUT
├── cloud_oauth.py      CloudOAuthMixin (auth / callback / manual / finish + helpers)
├── cloud_user.py       CloudUserMixin (email/password step)
├── modbus_only.py      ModbusOnlyMixin (local_setup / import / reconfigure_modbus)
├── zeroconf.py         ZeroconfMixin (WiNet-S mDNS discovery + confirm)
├── reconfigure.py      ReconfigureAndReauthMixin (reauth / reauth_confirm / reconfigure)
└── options.py          SungrowOptionsFlow
```

No behaviour changes — every step keeps its exact logic, docstring, and
signature. The refactor is purely structural.

Assembly
========

Each mixin subclasses `_SungrowFlowBase(ConfigFlow)` without passing
`domain=`, so none of them register as a config flow on their own — only
the concrete `SungrowConfigFlow(mixins..., domain=DOMAIN)` shell in
`__init__.py` does. `_SungrowFlowBase.__init__` owns the shared instance
state (init_info, auth_client, _state, _reauth_entry, ...) and
`async_remove` cleans up OAuth callback futures and state on flow
removal; every mixin inherits both.

Shared library imports
======================

To keep a single canonical patch target for tests, `_base.py` imports
`Auth`, `UserAuth`, `AuthError`, `PySolarCloudException`, and
`async_get_clientsession` and mixins access them via dotted lookup
(`_base.Auth`, `_base.async_get_clientsession(...)`). The original
monolithic file had one binding for each; the package preserves that
single-target shape via `_base` rather than duplicating from-imports
across four mixins that would each need their own patch site.

Tests updated accordingly:

* `config_flow.Auth`                    → `config_flow._base.Auth`
* `config_flow.UserAuth`                → `config_flow._base.UserAuth`
* `config_flow.async_get_clientsession` → `config_flow._base.async_get_clientsession`
* `config_flow.asyncio.wait_for`        → `config_flow.cloud_oauth.asyncio.wait_for`

Re-exports for external consumers
=================================

`config_flow/__init__.py` re-exports `SungrowConfigFlow`,
`SungrowOptionsFlow`, `_normalize_redirect_uri`, and
`_parse_winet_properties` at the package level, so tests and any
existing importers of `custom_components.sungrow.config_flow.<name>`
keep working unchanged.

Verification
============

* `ruff check custom_components/ tests/` — clean
* `ruff format --check custom_components/ tests/` — clean
* `mypy` (strict) — clean
* `pytest tests/` — 833 passed, 4 deselected (live)
