+++ tests/test_reconfigure_transport.py
@@ -25,7 +25,7 @@ from .conftest import MOCK_CONFIG_DATA
 def mock_client_session():
     """Mock async_get_clientsession."""
     with patch(
-        "custom_components.sungrow.config_flow.async_get_clientsession",
+        "custom_components.sungrow.config_flow._base.async_get_clientsession",
         return_value=MagicMock(),
     ):
         yield
